==> Transitions from smoking/7. vaping_plts_tabs.Rmd <==
---
title: "Plots and other tables"
author: "Alexandria Andrayas"
output: 
  html_document:
    toc: true
    toc_float:
      collapsed: false
      smooth_scroll: false
editor_options: 
  chunk_output_type: console
---
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

# Load packages

```{r load-packages}
library(ggplot2)
library(ggtext)
library(haven)
library(labelled)
library(sticky)
library(dplyr)
library(tidyr)
library(gtsummary)
library(ggpubr)
library(extrafont)
library(mice)
#library(tidycmprsk)
#library(ggsurvfit)
#font_import()
#loadfonts(device="win")
```

# Get data and specify labels

```{r data-labs}
# set up data
subdf <- read_dta('//rdsfcifs.acrc.bris.ac.uk/MRC-IEU-research/projects/ieu2/p7/025/working/data/vaping_transitions/vaping_sub_measures.dta')
subdf <- to_factor(subdf, sort_levels='auto')
subdf <- data.frame(subdf)
subdf <- sticky_all(subdf)
subdf$id <- 1:nrow(subdf)

# 4 baseline exposures
confounders <- c('sex', 'ethnicity', 'hhincome_11', 'parent_dailsmk_12')

# 11 main exposures
main_exp <- c('freqsmk_21', 'bingalc_20', 'cana_20', 'drug_20', 'friends_smk_20', 'qual_20', 'parent_21', 'townsend_21_gr', 'bmi_18_gr', 'exerc_18', 'mfq_21_gr')

# 5 timepoints
vaps <- c('vap22', 'vap23', 'vap24', 'vap28', 'vap30')

# Imputed data
load('//rdsfcifs.acrc.bris.ac.uk/MRC-IEU-research/projects/ieu2/p7/025/working/data/vaping_transitions/vaping_impdfs.rda')
impdfs <- lapply(impdfs, function(x){
  x[,'id'] <- 1:nrow(x)
  return(x)
})

load('//rdsfcifs.acrc.bris.ac.uk/MRC-IEU-research/projects/ieu2/p7/025/working/data/vaping_transitions/vaping_imp.rda')
long_impdf <- complete(imp, 'long')

# Convert firstevent to factor
long_impdf$firstevent <- factor(long_impdf$firstevent, levels=c(1:4), labels=c('No event','Dual use','Exclusive e-cigarette use','Neither use'))

# Proportions of first non-exclusive smoking events
round(prop.table(table(long_impdf$firstevent)),2)
round(prop.table(table(subdf$firstevent)),2)
```

# Descriptive stats tables

```{r desc-stats}
desc_tab <- tbl_summary(subdf[, c(confounders, main_exp)], statistic = all_continuous() ~ "{mean} ({sd})", type = all_dichotomous() ~ "categorical", missing = 'no') %>%
  bold_labels() %>%
  italicize_levels() %>%
  add_n(statistic = "{n_miss} ({p_miss}%)") %>%
  modify_header(n = "**Missing**")

desc_tab |> 
  as_gt() |> 
  gt::gtsave(filename = '//path/to/files/descstats_tab.docx')

subdf$completecase <- ifelse(apply(subdf[,c(confounders, main_exp)],1,anyNA)==F,1,0)
table(subdf$completecase)
#173
desc_tab <- tbl_summary(subdf[subdf$completecase %in% 1, c(confounders, main_exp)], statistic = all_continuous() ~ "{mean} ({sd})", type = all_dichotomous() ~ "categorical", missing = 'no') %>%
  bold_labels() %>%
  italicize_levels()

desc_tab |> 
  as_gt() |> 
  gt::gtsave(filename = '//path/to/files/descstats_cc_tab.docx')

desc_tabs_byoutcomes <- tbl_merge(tbls = list(
  tbl_summary(subdf[, c(confounders, main_exp, 'nonuse')], by=nonuse, statistic=all_continuous() ~ "{mean} ({sd})", type = all_dichotomous() ~ "categorical", missing = 'no') %>% add_p(),
  tbl_summary(subdf[, c(confounders, main_exp, 'exclecig')], by=exclecig, statistic=all_continuous() ~ "{mean} ({sd})", type = all_dichotomous() ~ "categorical", missing = 'no') %>% add_p(),
  tbl_summary(subdf[, c(confounders, main_exp, 'dual')], by=dual, statistic=all_continuous() ~ "{mean} ({sd})", type = all_dichotomous() ~ "categorical", missing = 'no') %>% add_p(),
  tbl_summary(subdf[, c(confounders, main_exp, 'ecig')], by=ecig, statistic=all_continuous() ~ "{mean} ({sd})", type = all_dichotomous() ~ "categorical", missing = 'no') %>% add_p(),
  tbl_summary(subdf[!is.na(subdf$firstevent), c(confounders, main_exp)], statistic=all_continuous() ~ "{mean} ({sd})", type = all_dichotomous() ~ "categorical", missing = 'no')
  ), tab_spanner = c('**Neither use**','**Exclusive e-cig use**','**Dual use**','**Any e-cig use**','**Overall**')) %>%
  bold_labels() %>%
  italicize_levels()

desc_tabs_byoutcomes |> 
  as_gt() |> 
  gt::gtsave(filename = '//path/to/files/descstats_byoutcomes_tab.docx')

#check differences in main exposures by early-life confounders
desc_tabs_byconfounders <- tbl_merge(tbls = list(
  tbl_summary(subdf[, c(confounders, main_exp)], by=sex, statistic=all_continuous() ~ "{mean} ({sd})", type = all_dichotomous() ~ "categorical", missing = 'no') %>% add_p(),
  tbl_summary(subdf[, c(confounders, main_exp)], by=ethnicity, statistic=all_continuous() ~ "{mean} ({sd})", type = all_dichotomous() ~ "categorical", missing = 'no') %>% add_p(),
  tbl_summary(subdf[, c(confounders, main_exp)], by=hhincome_11, statistic=all_continuous() ~ "{mean} ({sd})", type = all_dichotomous() ~ "categorical", missing = 'no') %>% add_p(),
  tbl_summary(subdf[, c(confounders, main_exp)], by=parent_dailsmk_12, statistic=all_continuous() ~ "{mean} ({sd})", type = all_dichotomous() ~ "categorical", missing = 'no') %>% add_p()
  ), tab_spanner = c('**Sex assigned at birth**','**Ethnicity**','**Household income (11y)**','**Parental smoking (12y)**')) %>%
  bold_labels() %>%
  italicize_levels()

desc_tabs_byconfounders |> 
  as_gt() |> 
  gt::gtsave(filename = '//path/to/files/descstats_byconfounders_tab.docx')

subdf$censored[subdf$losfoll %in% 1] <- 1
desc_tabs_censored <- tbl_summary(subdf[, c(confounders, main_exp, 'censored')], by=censored, statistic=all_continuous() ~ "{mean} ({sd})", type = all_dichotomous() ~ "categorical", missing = 'no') %>% add_p()

desc_tabs_censored |> 
  as_gt() |> 
  gt::gtsave(filename = '//path/to/files/descstats_censored_tab.docx')
```

# Box plots showing nicotine use categories and missingness per timepoint

```{r box-plts}
# Initialise list for plots
p <- list()

# Number of people per nicotine use category by timepoint (complete cases)
ns <- data.frame(n = as.vector(sapply(subdf[,vaps], table, useNA = "ifany")), 
                 cat = rep(c('Exclusive smoking','Dual use','Exclusive e-cig use','No smoking nor e-cig use','Missing'), 5),
                 time = rep(c(22,23,24,28,30), each=5))
ns$cat <- factor(ns$cat, levels=c('Exclusive smoking','Dual use','Exclusive e-cig use','No smoking nor e-cig use','Missing'))

# Plot
p$ncat <- ggplot(ns, aes(x = time, y = n, fill = cat)) +
  geom_bar(stat="identity", position = 'dodge', colour = 'black') +
  labs(x = '\nTimepoint', y = 'N\n', fill = '') +
  theme_bw() +
  theme(plot.title = element_text(face="bold", hjust=0.5), 
        legend.position = 'bottom', 
        legend.spacing.x = unit(5, 'cm'), 
        text=element_text(size=20, family = 'Times New Roman')) +
  scale_x_continuous(breaks=seq(22,30,1)) +
  scale_y_continuous(limits=c(0,450), breaks=seq(0,450,50)) +
  scale_fill_manual(values=c("grey30", "grey50", "grey70", "grey90", "white")) +
  guides(fill = guide_legend(byrow = TRUE), colour = 'none')

# Save
ggsave(p$ncat, width=12,   height=10, filename = '//path/to/files/ns_percat_plt.png')
```

# Cumulative incidence plots

```{r cuminc-plts}
# Fit survival curve 
fit <- survival::survfit(survival::Surv(timeto, firstevent) ~ 1, data = long_impdf)

# Pull data and tidy
psta <- as.data.frame(fit$pstate)
colnames(psta) <- fit$states
psta$times <- fit$time
event <- value <- strata <- NULL
pstal <- gather(psta, event, value, -times)
pstal$event[pstal$event %in% '(s0)'] <- 'Exclusive smoking'
pstal$event[pstal$event %in% 'Neither use'] <- 'No nicotine use'
pstal <- rbind(pstal,
               data.frame(times=rep(0,4),
                          event=c('Exclusive smoking','No nicotine use',
                                  'Exclusive e-cigarette use','Dual use'),
                          value=c(1,0,0,0)))
pstal <- pstal[order(pstal$event, pstal$times, pstal$value),]
pstal$event <- factor(pstal$event, 
                      levels=c('Exclusive smoking','No nicotine use',
                               'Exclusive e-cigarette use','Dual use'))

# Plot
p$cuminc_imp <- ggplot(pstal, aes(times, value, fill=event)) +
  geom_area() + 
  scale_x_continuous(limits=c(0,9), breaks=c(0,1,2,3,7,9)) +
  theme_bw() +
  theme(text=element_text(size = 20, family = 'Aptos'), 
        legend.position = 'bottom', legend.box='vertical', 
        legend.key=element_rect(colour="black")) +
  labs(x = '\nApproximate years from 21+ questionnaire\n', y = 'Probability\n', fill = NULL) +
  scale_fill_manual(values=c('Exclusive smoking'='#B3C3CB', 'No nicotine use'='#83949B', 
                             'Exclusive e-cigarette use'='#596369', 'Dual use'='#323D42'))

# Save
ggsave(p$cuminc_imp, width=12, height=10, dpi = 800, filename = '//path/to/files/cuminc_stacked_imp_plt.png')
```

# Plots with subdistribution hazard results

```{r produce-surv-plots}
# Load model results
load('//path/to/files/subdistfigdat_adj.rda')

# Change outcome labels
levels(subdistfigdat$imp_w$outcome)[levels(subdistfigdat$imp_w$outcome) %in% 'E-cigarette use'] <- 'Any e-cigarette use'
levels(subdistfigdat$imp_w$outcome)[levels(subdistfigdat$imp_w$outcome) %in% 'Exclusive e-cig use'] <- 'Exclusive e-cigarette use'
levels(subdistfigdat$imp_w$outcome)[levels(subdistfigdat$imp_w$outcome) %in% 'Non-use'] <- 'No nicotine use'

# Remove duplicate rows
subdistfigdat$imp_w <- subdistfigdat$imp_w[!(subdistfigdat$imp_w$outcome %in% 'No nicotine use' & subdistfigdat$imp_w$ecig_collapsed %in% T),]

# Change figure labels
subdistfigdat$imp_w$var_label <- recode(subdistfigdat$imp_w$var_label, 
  "Smoking frequency (21y)" = "Higher smoking frequency<sup><i>a</i></sup> (21y)",
  "Binge drinking, six or more units (20y)" = "Higher binge drinking frequency<sup><i>b</i></sup> (20y)",
  "Cannabis use (20y)" = "Higher cannabis use<sup><i>c</i></sup> (20y)",
  "Drug use (20y)" = "Drug use in past year<sup><i>d</i></sup> (20y)",
  "Peer smoking (20y)" = "More peer smoking<sup><i>e</i></sup> (20y)",
  "Educational qualifications (20y)" = "Lower education<sup><i>f</i></sup> (20y)",
  "Is a parent (21y)" = "Early parenthood<sup><i>g</i></sup> (21y)",
  "G1 Townsend deprivation score, quintiles (21y)" = "Higher neighbourhood deprivation<sup><i>h</i></sup> (21y)",
  "BMI (18y)" = "Higher BMI<sup><i>i</i></sup> (18y)",
  "Frequency of exercise, past year (18y)" = "Lower exercise frequency<sup><i>j</i></sup> (18y)",
  "Mood and Feelings Questionnaire scores (21y)" = "More depressive symptoms<sup><i>k</i></sup> (21y)"
  )

# Change labels to factor
subdistfigdat$imp_w$var_label <- factor(subdistfigdat$imp_w$var_label, levels=c(
  "Sex assigned at birth", 
  "Ethnicity", 
  "Household income (11y)", 
  "Parental smoking (12y)", 
  "Higher smoking frequency<sup><i>a</i></sup> (21y)",
  "Higher binge drinking frequency<sup><i>b</i></sup> (20y)",
  "Higher cannabis use<sup><i>c</i></sup> (20y)",
  "Drug use in past year<sup><i>d</i></sup> (20y)",
  "More peer smoking<sup><i>e</i></sup> (20y)",
  "Lower education<sup><i>f</i></sup> (20y)",
  "Early parenthood<sup><i>g</i></sup> (21y)",
  "Higher neighbourhood deprivation<sup><i>h</i></sup> (21y)",
  "Higher BMI<sup><i>i</i></sup> (18y)",
  "Lower exercise frequency<sup><i>j</i></sup> (18y)",
  "More depressive symptoms<sup><i>k</i></sup> (21y)"))

# Set colours
cols <- c("#B61516","#C9665B","#C97B7A","#CA904E","#A99844",
          "#71A99F","#85A5B6","#196090","#668EA9","#9895AE","#BC9AAB")
names(cols) <- levels(subdistfigdat$imp_w$var_label)[5:15]
subdistfigdat$imp_w$cols <- recode(subdistfigdat$imp_w$var_label, !!!cols)

# Create variable to map labels
subdistfigdat$imp_w$y <- interaction(subdistfigdat$imp_w$var_label, subdistfigdat$imp_w$label)
subdistfigdat$imp_w$y <- factor(subdistfigdat$imp_w$y, levels=unique(subdistfigdat$imp_w$y))

# Remove any e-cig use outcome
subdistfigdat$imp_w <- subdistfigdat$imp_w[!subdistfigdat$imp_w$outcome %in% 'Any e-cigarette use',]

# Plot function
pltit <- function(df){
  df <- df[!df$var %in% c('sex','ethnicity','hhincome_11','parent_dailsmk_12'),]
  df <- df[!df$ref %in% T,]
  pd <- position_dodge(0.8)
  p <- ggplot(df, aes(x=HR, xmin=LCI, xmax=UCI, y=y, fill=outcome)) +
    geom_vline(xintercept = 1, color = "grey40", linetype = "dashed") +
    geom_errorbar(width=.3, position = pd, colour='black') +
    geom_point(size=5, shape=21, position = pd, colour='black') +
    facet_grid(var_label ~ outcome, scales = "free_y", switch = 'y') +
    scale_x_log10() + 
    scale_y_discrete(labels=setNames(df$label, df$y), position = "right") +
    scale_linetype_manual(values=1) +
    #scale_colour_identity() +
    scale_fill_manual(values=c('No nicotine use'='#83949B', 
                                 'Exclusive e-cigarette use'='#596369', 
                                 'Dual use'='#323D42')) +
    theme_bw() +
    theme(legend.position='none',
          text=element_text(size=30, family = 'Aptos'), 
          strip.background = element_blank(), 
          strip.placement = "outside", 
          strip.text.x = element_text(face = 'bold'),
          strip.text.y.left = element_markdown(angle = 0, hjust = 1),
          axis.title.x = element_text(size = 25),
          axis.text.y=element_blank()) +
    labs(x = "\nSubdistribution hazard ratio (SHR)\n", y = '')
  return(p)
}

# Get plots
subdistplts <- lapply(subdistfigdat, pltit)

# Save
ggsave(subdistplts$imp_w, width=26, height=18, dpi = 800, filename = '//path/to/files/subdist_imp_w_plt_nolabs.png')
```
